fix/validate route parameters

+++ app/Http/Controllers/CamisetaController.php
@@ -26,27 +26,26 @@ class CamisetaController extends Controller
      */
     public function indexByCategory(string $id)
     {
-        try {
-            $validatedData = Validator::make(['id' => $id], ['id' => 'integer',])->validate();
 
-            $categoria = Categoria::find($id);
-            if ($categoria == null) {
-                abort(404);
-            }
-
-            $camisetas = $categoria->camisetas;
-            $camisetas = $this->loadImages($camisetas);
-            return view('tables.camisetas', ['camisetas' => $camisetas, 'categoria' => $categoria]);
-        } catch (ValidationException $e) {
+        $validator = Validator::make(['id' => $id], ['id' => 'integer',]);
+        if ($validator->fails())
+            abort(404);
+        $categoria = Categoria::find($id);
+        if ($categoria == null) {
             abort(404);
         }
+
+        $camisetas = $categoria->camisetas;
+        $camisetas = $this->loadImages($camisetas);
+        return view('tables.camisetas', ['camisetas' => $camisetas, 'categoria' => $categoria]);
+
     }
 
     public function getCamisetasAPI()
     {
         $camisetas = Camiseta::all();
         $camisetas = $this->loadImages($camisetas);
-        
+
         return response()->json($camisetas);
     }
 
@@ -54,19 +53,18 @@ class CamisetaController extends Controller
     {
         $validator = Validator::make(['id' => $id], ['id' => 'integer',]);
 
-        if ($validator->fails())
-        {
+        if ($validator->fails()) {
             return response('El ID de la categoría tiene que ser un valor entero', 400);
         }
 
         $categoria = Categoria::find($id);
         if ($categoria == null) {
-            return response('No existe categoría con ID '.$id, 404);
+            return response('No existe categoría con ID ' . $id, 404);
         }
 
         $camisetas = $categoria->camisetas();
         $camisetas = $this->loadImages($camisetas);
-        
+
         return response()->json($camisetas);
     }
 
@@ -85,7 +83,7 @@ class CamisetaController extends Controller
                 fwrite($file, $image);
                 fclose($file);
             }
-            $camiseta->imagen_frente = $image_route;
+            $camiseta->imagen_frente = "/" . $image_route;
 
             $image_route = "images/" . $camiseta->id . "atras_" . $updatedDate . ".jpg";
             if (!file_exists($image_route)) {
@@ -94,7 +92,7 @@ class CamisetaController extends Controller
                 fwrite($file, $image);
                 fclose($file);
             }
-            $camiseta->imagen_atras = $image_route;
+            $camiseta->imagen_atras = "/" . $image_route;
         }
         return $camisetas;
     }
@@ -113,32 +111,33 @@ class CamisetaController extends Controller
      */
     public function store(Request $request)
     {
-        $request->validate([
-            'nombre' => ['required', 'regex:/^[a-zA-Z0-9\s\/]+$/', 'unique:camisetas,nombre'],
-            'descripcion' => ["required", 'regex:/^[a-zA-Z0-9\s\/]+$/'],
-            'precio' => ["required", 'decimal:0,2'],
-            'talles' => ["required","array","min:1"], 
-            'tags' => ["required","array","min:1"], 
-            'tags.*' => ['regex:/^[a-zA-Z\s]+$/'], 
-            'imagen_frente' => ["required", "image"],
-            'imagen_atras' => ["required", "image"],
-        ],
-        [
-            'nombre.required' => 'Este campo no puede estar vacío',
-            'nombre.regex' => 'Este campo no puedo contener caracteres especiales',
-            'nombre.unique' => 'No puede haber camisetas con nombres repetidos',
-            'descripcion.required' => 'Este campo no puede estar vacío',
-            'descripcion.regex' => 'Este campo no puedo contener caracteres especiales',
-            'precio.required' => 'Este campo no puede estar vacío',
-            'precio' => 'El precio debe ser un numero decimal, con maximo 2 cifras de centavos',
-            'talles.required' => 'La camiseta debe tener al menos un talle',
-            'tags.required' => 'Completar con al menos una categoría',
-            'tags.*.regex' => 'Una categoría no puede tener caracteres especiales o números',
-            'imagen_frente.required' => 'Una camiseta necesita una imágen de frente',
-            'imagen_atras.required' => 'Una camiseta necesita una imágen de atrás',
-            'imagen_frente.image' => 'La imágen subida debe estar en un formato permitido: jpg, png, webp',
-            'imagen_atras.image' => 'La imágen subida debe estar en un formato permitido: jpg, png, webp',
-        ]
+        $request->validate(
+            [
+                'nombre' => ['required', 'regex:/^[a-zA-Z0-9\s\/]+$/', 'unique:camisetas,nombre'],
+                'descripcion' => ["required", 'regex:/^[a-zA-Z0-9\s\/]+$/'],
+                'precio' => ["required", 'decimal:0,2'],
+                'talles' => ["required", "array", "min:1"],
+                'tags' => ["required", "array", "min:1"],
+                'tags.*' => ['regex:/^[a-zA-Z\s]+$/'],
+                'imagen_frente' => ["required", "image"],
+                'imagen_atras' => ["required", "image"],
+            ],
+            [
+                'nombre.required' => 'Este campo no puede estar vacío',
+                'nombre.regex' => 'Este campo no puedo contener caracteres especiales',
+                'nombre.unique' => 'No puede haber camisetas con nombres repetidos',
+                'descripcion.required' => 'Este campo no puede estar vacío',
+                'descripcion.regex' => 'Este campo no puedo contener caracteres especiales',
+                'precio.required' => 'Este campo no puede estar vacío',
+                'precio' => 'El precio debe ser un numero decimal, con maximo 2 cifras de centavos',
+                'talles.required' => 'La camiseta debe tener al menos un talle',
+                'tags.required' => 'Completar con al menos una categoría',
+                'tags.*.regex' => 'Una categoría no puede tener caracteres especiales o números',
+                'imagen_frente.required' => 'Una camiseta necesita una imágen de frente',
+                'imagen_atras.required' => 'Una camiseta necesita una imágen de atrás',
+                'imagen_frente.image' => 'La imágen subida debe estar en un formato permitido: jpg, png, webp',
+                'imagen_atras.image' => 'La imágen subida debe estar en un formato permitido: jpg, png, webp',
+            ]
         );
 
         $camiseta = new Camiseta();
@@ -148,8 +147,8 @@ class CamisetaController extends Controller
         $camiseta->precio = $request->get('precio');
 
         $talles = "";
-        foreach($request->get('talles') as $talle){
-            $talles .= $talle.", ";
+        foreach ($request->get('talles') as $talle) {
+            $talles .= $talle . ", ";
         }
         $talles = substr($talles, 0, -2);
         $camiseta->talles_disponibles = $talles;
@@ -170,7 +169,7 @@ class CamisetaController extends Controller
         $id_categorias = array();
 
         foreach ($categorias as $categoria) {
-            $existe = Categoria::where('name',$categoria)->first();
+            $existe = Categoria::where('name', $categoria)->first();
             if ($existe != null) {
                 array_push($id_categorias, $existe);
             } else {
@@ -201,21 +200,19 @@ class CamisetaController extends Controller
      */
     public function edit(string $id)
     {
-        try {
-            $validatedData = Validator::make(['id' => $id], ['id' => 'integer',])->validate();
-
-            $camiseta = Camiseta::find($id);
-
-            if ($camiseta == null) {
-                abort(404);
-            }
-
-            $categorias = Categoria::all()->pluck('name');
-
-            return view('edit.edit_camiseta', ['camiseta' => $camiseta, 'categorias' => $categorias]);
-        } catch (ValidationException $e) {
+        $validator = Validator::make(['id' => $id], ['id' => 'integer',]);
+        if ($validator->fails())
             abort(404);
-        }        
+        $camiseta = Camiseta::find($id);
+
+        if ($camiseta == null) {
+            abort(404);
+        }
+
+        $categorias = Categoria::all()->pluck('name');
+
+        return view('edit.edit_camiseta', ['camiseta' => $camiseta, 'categorias' => $categorias]);
+
     }
 
     /**
@@ -223,41 +220,40 @@ class CamisetaController extends Controller
      */
     public function update(Request $request, string $id)
     {
-        $camiseta = null; // valida id
-        try {
-            $validatedData = Validator::make(['id' => $id], ['id' => 'integer',])->validate();
-            $camiseta = Camiseta::find($id);
-            if ($camiseta == null) {
-                abort(404);
-            }
-        } catch (ValidationException $e) {
+        $validator = Validator::make(['id' => $id], ['id' => 'integer',]);
+        if ($validator->fails())
             abort(404);
-        }        
+        $camiseta = Camiseta::find($id);
+        if ($camiseta == null) {
+            abort(404);
+        }
+
         // valida campos form
-        $request->validate([
-            'nombre' => ['required', 'regex:/^[a-zA-Z0-9\s\/]+$/', Rule::unique('camisetas','nombre')->ignore($id)],
-            'descripcion' => ["required", 'regex:/^[a-zA-Z0-9\s\/]+$/'],
-            'precio' => ["required", 'decimal:0,2'],
-            'talles' => ["required","array","min:1"], 
-            'tags' => ["required","array","min:1"], 
-            'tags.*' => ['regex:/^[a-zA-Z0-9\s]+$/'], 
-            'imagen_atras' => ["nullable", "image"],
-            'imagen_frente' => ["nullable", "image"],
-        ],
-        [
-            'nombre.required' => 'Este campo no puede estar vacío',
-            'nombre.regex' => 'Este campo no puedo contener caracteres especiales',
-            'nombre.unique' => 'No puede haber camisetas con nombres repetidos',
-            'descripcion.required' => 'Este campo no puede estar vacío',
-            'descripcion.regex' => 'Este campo no puedo contener caracteres especiales',
-            'precio.required' => 'Este campo no puede estar vacío',
-            'precio' => 'El precio debe ser un numero decimal, con maximo 2 cifras de centavos',
-            'talles.required' => 'La camiseta debe tener al menos un talle',
-            'tags.required' => 'Completar con al menos una categoría',
-            'tags.*.regex' => 'Una categoría no puede tener caracteres especiales o números',
-            'imagen_atras.image' => 'La imágen subida debe estar en un formato permitido: jpg, png, webp',
-            'imagen_frente.image' => 'La imágen subida debe estar en un formato permitido: jpg, png, webp',
-        ]
+        $request->validate(
+            [
+                'nombre' => ['required', 'regex:/^[a-zA-Z0-9\s\/]+$/', Rule::unique('camisetas', 'nombre')->ignore($id)],
+                'descripcion' => ["required", 'regex:/^[a-zA-Z0-9\s\/]+$/'],
+                'precio' => ["required", 'decimal:0,2'],
+                'talles' => ["required", "array", "min:1"],
+                'tags' => ["required", "array", "min:1"],
+                'tags.*' => ['regex:/^[a-zA-Z0-9\s]+$/'],
+                'imagen_atras' => ["nullable", "image"],
+                'imagen_frente' => ["nullable", "image"],
+            ],
+            [
+                'nombre.required' => 'Este campo no puede estar vacío',
+                'nombre.regex' => 'Este campo no puedo contener caracteres especiales',
+                'nombre.unique' => 'No puede haber camisetas con nombres repetidos',
+                'descripcion.required' => 'Este campo no puede estar vacío',
+                'descripcion.regex' => 'Este campo no puedo contener caracteres especiales',
+                'precio.required' => 'Este campo no puede estar vacío',
+                'precio' => 'El precio debe ser un numero decimal, con maximo 2 cifras de centavos',
+                'talles.required' => 'La camiseta debe tener al menos un talle',
+                'tags.required' => 'Completar con al menos una categoría',
+                'tags.*.regex' => 'Una categoría no puede tener caracteres especiales o números',
+                'imagen_atras.image' => 'La imágen subida debe estar en un formato permitido: jpg, png, webp',
+                'imagen_frente.image' => 'La imágen subida debe estar en un formato permitido: jpg, png, webp',
+            ]
         );
         $camiseta = Camiseta::where('id', $id)->first();
 
@@ -266,14 +262,14 @@ class CamisetaController extends Controller
         $camiseta->precio = $request->get('precio');
 
         $talles = "";
-        foreach($request->get('talles') as $talle){
-            $talles .= $talle.", ";
+        foreach ($request->get('talles') as $talle) {
+            $talles .= $talle . ", ";
         }
         $talles = substr($talles, 0, -2);
         $camiseta->talles_disponibles = $talles;
 
-        $this->unlinkOldImages($camiseta->id,$camiseta->updated_at);
-        
+        $this->unlinkOldImages($camiseta->id, $camiseta->updated_at);
+
         $atras = $request->file('imagen_atras');
         $frente = $request->file('imagen_frente');
         if ($atras != null) {
@@ -284,15 +280,15 @@ class CamisetaController extends Controller
             $frente = base64_encode($frente->get());
             $camiseta->imagen_frente = $frente;
         }
-        
+
         $camiseta->update();
 
-        $camiseta->categorias()->detach(); 
+        $camiseta->categorias()->detach();
 
         $categorias = $request->get('tags');
         $id_categorias = array();
         foreach ($categorias as $categoria) {
-            $existe = Categoria::where('name',$categoria)->first();
+            $existe = Categoria::where('name', $categoria)->first();
             if ($existe != null) {
                 array_push($id_categorias, $existe);
             } else {
@@ -305,18 +301,18 @@ class CamisetaController extends Controller
 
         foreach ($id_categorias as $id_categoria) {
             $camiseta->categorias()->attach($id_categoria);
-        } 
+        }
 
         return redirect('/camisetas')->with("success", "La camiseta '" . $camiseta->nombre . "' fue actualizada con éxito");
     }
 
     private function unlinkOldImages($camisetaID, $unformattedDate)
     {
-        $updatedDate = str_replace(' ', '_',str_replace(":", "-", $unformattedDate));
-        $imageRoute = "images/".$camisetaID."atras_".$updatedDate.".jpg";
-        unlink($imageRoute); 
-        $imageRoute = "images/".$camisetaID."frente_".$updatedDate.".jpg";
-        unlink($imageRoute); 
+        $updatedDate = str_replace(' ', '_', str_replace(":", "-", $unformattedDate));
+        $imageRoute = "images/" . $camisetaID . "atras_" . $updatedDate . ".jpg";
+        unlink($imageRoute);
+        $imageRoute = "images/" . $camisetaID . "frente_" . $updatedDate . ".jpg";
+        unlink($imageRoute);
     }
 
     /**
@@ -324,28 +320,26 @@ class CamisetaController extends Controller
      */
     public function updateStock(string $id)
     {
-        try {
-            $validatedData = Validator::make(['id' => $id], ['id' => 'integer',])->validate();
+        $validator = Validator::make(['id' => $id], ['id' => 'integer',]);
+        if ($validator->fails())
+            abort(404);
+        $camiseta = Camiseta::find($id);
 
-            $camiseta = Camiseta::find($id);
-
-            if ($camiseta == null) {
-                abort(404);
-            }
-
-            $nuevo = 0;
-            if ($camiseta->activo == 0)
-                $nuevo = 1;
-
-            $this->unlinkOldImages($camiseta->id,$camiseta->updated_at);
-
-            $camiseta->activo = $nuevo;
-            $camiseta->update();
-
-            return redirect()->back()->with("success", "El stock de la camiseta '" . $camiseta->nombre . "' fue actualizado con éxito");
-        } catch (ValidationException $e) {
+        if ($camiseta == null) {
             abort(404);
         }
+
+        $nuevo = 0;
+        if ($camiseta->activo == 0)
+            $nuevo = 1;
+
+        $this->unlinkOldImages($camiseta->id, $camiseta->updated_at);
+
+        $camiseta->activo = $nuevo;
+        $camiseta->update();
+
+        return redirect()->back()->with("success", "El stock de la camiseta '" . $camiseta->nombre . "' fue actualizado con éxito");
+
     }
 
     /**
@@ -353,18 +347,16 @@ class CamisetaController extends Controller
      */
     public function delete(string $id)
     {
-        try {
-            $validatedData = Validator::make(['id' => $id], ['id' => 'integer',])->validate();
-
-            $camiseta = Camiseta::find($id);
-            if ($camiseta == null) {
-                abort(404);
-            }
-
-            return view('edit.delete_camiseta', ['camiseta' => $camiseta]);
-        } catch (ValidationException $e) {
+        $validator = Validator::make(['id' => $id], ['id' => 'integer',]);
+        if ($validator->fails())
+            abort(404);
+        $camiseta = Camiseta::find($id);
+        if ($camiseta == null) {
             abort(404);
         }
+
+        return view('edit.delete_camiseta', ['camiseta' => $camiseta]);
+
     }
 
     /**
@@ -372,21 +364,19 @@ class CamisetaController extends Controller
      */
     public function destroy(string $id)
     {
-        try {
-            $validatedData = Validator::make(['id' => $id], ['id' => 'integer',])->validate();
+        $validator = Validator::make(['id' => $id], ['id' => 'integer',]);
+        if ($validator->fails())
+            abort(404);
+        $camiseta = Camiseta::find($id);
 
-            $camiseta = Camiseta::find($id);
 
-
-            if ($camiseta == null) {
-                abort(404);
-            }
-
-            $nombre = $camiseta->nombre;
-            $camiseta->delete();
-            return redirect('/camisetas')->with("deleted", "La camiseta '" . $nombre . "' fue eliminada con éxito");
-        } catch (ValidationException $e) {
+        if ($camiseta == null) {
             abort(404);
         }
+
+        $nombre = $camiseta->nombre;
+        $camiseta->delete();
+        return redirect('/camisetas')->with("deleted", "La camiseta '" . $nombre . "' fue eliminada con éxito");
+
     }
 }
