+++ app/Http/Controllers/CategoriaController.php
@@ -36,14 +36,15 @@ class CategoriaController extends Controller
      */
     public function store(Request $request)
     {
-        $request->validate([
-            'nombre' => ['required', 'regex:/^[a-zA-Z\s]+$/', 'unique:categorias,name'],
-        ],
-        [
-            'nombre.required' => 'Una categoría debe tener nombre',
-            'nombre.regex' => 'Una categoría no puede tener caracteres especiales ni números',
-            'nombre.unique' => 'Ya existe una categoría con este nombre',
-        ]
+        $request->validate(
+            [
+                'nombre' => ['required', 'regex:/^[a-zA-Z\s]+$/', 'unique:categorias,name'],
+            ],
+            [
+                'nombre.required' => 'Una categoría debe tener nombre',
+                'nombre.regex' => 'Una categoría no puede tener caracteres especiales ni números',
+                'nombre.unique' => 'Ya existe una categoría con este nombre',
+            ]
         );
 
         $newCategoria = new Categoria();
@@ -51,7 +52,7 @@ class CategoriaController extends Controller
 
         $newCategoria->save();
 
-        return redirect('/categorias')->with('success', "La categoría '".$newCategoria->name."' fue creada con éxito");
+        return redirect('/categorias')->with('success', "La categoría '" . $newCategoria->name . "' fue creada con éxito");
     }
 
     /**
@@ -83,18 +84,15 @@ class CategoriaController extends Controller
      */
     public function delete(string $id)
     {
-        try {
-            $validatedData = Validator::make(['id' => $id], ['id' => 'integer',])->validate();
-
-            $categoria = Categoria::find($id);
-            if($categoria == null){
-                abort(404);
-            }
-            
-            return view('edit.delete_categoria', ['categoria' => $categoria]);
-        } catch (ValidationException $e) {
+        $validator = Validator::make(['id' => $id], ['id' => 'integer',]);
+        if ($validator->fails())
+            abort(404);
+        $categoria = Categoria::find($id);
+        if ($categoria == null) {
             abort(404);
         }
+
+        return view('edit.delete_categoria', ['categoria' => $categoria]);
     }
 
     /**
@@ -102,21 +100,19 @@ class CategoriaController extends Controller
      */
     public function destroy(string $id)
     {
-        try {
-            $validatedData = Validator::make(['id' => $id], ['id' => 'integer',])->validate();
-
-            $categoria = Categoria::find($id);
-            if($categoria == null){
-                abort(404);
-            }
-            $nombre = $categoria->name;
-
-            $categoria->camisetas()->detach(); // Borra la relacion camiseta_categoria para todas las camisetas asociadas a esta categoria
-            $categoria->delete();
-            
-            return redirect('/categorias')->with("deleted", "La categoría ".$nombre.' fue eliminada con éxito');
-        } catch (ValidationException $e) {
+        $validator = Validator::make(['id' => $id], ['id' => 'integer',]);
+        if ($validator->fails())
+            abort(404);
+        $categoria = Categoria::find($id);
+        if ($categoria == null) {
             abort(404);
         }
+        $nombre = $categoria->name;
+
+        $categoria->camisetas()->detach(); // Borra la relacion camiseta_categoria para todas las camisetas asociadas a esta categoria
+        $categoria->delete();
+
+        return redirect('/categorias')->with("deleted", "La categoría " . $nombre . ' fue eliminada con éxito');
+
     }
 }
